Update patch set 1
Patch Set 1:

> I don't know what the real problem is.  I would want to see an
 > example of the problem using the go tool and GOPATH, not direct
 > tool invocations from the test directory.
The real problem is that local imports of the form "./p" is not
supported on windows. The problem will not affect packages in
GOROOT or GOPATH.

bug302 doesn't need to actually import the package p using
local import, so my CL switched to use -I . and regular import "p".

Patch-set: 1