Update patch set 1
Patch Set 1:

Part of the reason this was commented out is that I am not sure that it works. Can you

go build -a -gcflags -S runtime >x 2>&1

and double-check that there are DATA lines for runtime.earlycgocallback? It needs to be linker-initialized, not startup-initialized, because startup hasn't happened yet.

Patch-set: 1