Update patch set 2
Patch Set 2:

Thanks Dave for the review.  In my first version of this change - see Patch Set 2 - I added a new function RawEncoding() instead of adding a variadic options argument to NewEncoding(), but I was dissatisfied with that approach because it seemed to be going down the Win32 API's garden path of proliferating API variants.  I'm happy to go back to that approach if that's the right way, however.

Would you mind clarifying how adding a variadic options-list like this breaks the Go1 contract, given that existing code will still be able to call NewEncoding(str) without any options according to the original signature (i.e., with a zero-length options list)?  Are you concerned about impact on code that might notice the change by reflection on the NewEncoding method, or impact on other code-management tools?

Thanks
Bryan

Patch-set: 2
Label: Code-Review=0