Update patch set 5
Patch Set 5:

> My only minor worry with the SetPadding() approach is if users
 > accidentally call SetPadding() on the system-default Encoding
 > objects, causing confusion when the modification mysteriously leaks
 > through to other parts of the system.  This risk would not be
 > remotely unique to this package, however.  Just in case, I added a
 > simple sanity-check to SetPadding.
Good point. How about making SetPadding returns a new Encoding
instead of modifying the encoding in place?

I guess in that case, we need to change the name of the method to
emphasize that it doesn't change the underlying encoding.

We also need to consider what if we need to introduce new options
later.
enc := base64.StdEncoding.SetPadding(somePadding).SetAnotherOption(someValue)

Although the new proposal allocates needlessly for the first setting (padding),
it does look better and encoding is normally created at program startup,
so I don't think the extra allocation is going to hurt.

Let's settle on the API design first and then do the coding.

Patch-set: 5
Label: Code-Review=0