Update patch set 6
Patch Set 6:

True.  But unlike encoding/base64, encoding/xml and encoding/json do not provide variables representing "standard" Encoders that are expected to be constant; the caller must create her own xml or json Encoder before using it, so there's no risk of the caller accidentally modifying such a "standard" encoder.  In the case of encoding/base64, I expect almost all callers will use the standard encoders, and invoking NewEncoder() will be a rare exception, since there are just not that many different flavors of base64 out there (and with the proposed patch most of them will be covered by the set of standard encoders provided).  Thus, I would suggest that the fact that this risk is present in base64 but not xml or json might justify the inconsistency with the other encoding APIs.

But I have no strong feelings one way or another about this; I'm happy to go with whatever approach the Go Gods prefer. :)

B

Patch-set: 6
Label: Code-Review=0