os: Reduce allocations made by the os package
The commonly used os.(*File).Write is allocating
two chunks of 8 bytes because it calls runtime.convT2I
twice. This change adds a test to detect that this is
happening. The test is currently failing.

The proposed fix (at the end of error.go) is currently
disabled. When enabled, the test stops failing. The
new predicate errEq is correctly inlined by 6g, so in
addition to making less garbage, this code potentially
runs slightly faster as well.

Fixes #9370.

Change-Id: I363dbb1b956c40baef650b49e04b9a524964bee1