Update patch set 1
Patch Set 1:

> Ok I'll do that. But what the point exactly?
 > Do you want two internal packages, one in mime/multipart and one in
 > net/mail?

No, we don't need two internal packages. I actually don't care whether we have 1 or 7... they're internal so they don't affect the public-facing API. Whatever works.

But what I do care about is that the git history looks nice and that the moves/renames show up, especially the tests. There have been years of bug reports & fixes to this code... we're not going to throw it all out and rewrite it all. If I'm going to have any confidence in this move (and subsequent exporting of a new quoted printable package), it needs to be obvious it's just a move and not a rewrite.

Patch-set: 1
Label: Code-Review=0