Update patch set 4
Patch Set 4:

This is pretty awful. I would feel a lot better if we understood exactly when /dev/urandom can return EAGAIN. It's not supposed to, not in the way we use it.

If there is a general problem with files returning EAGAIN despite not being in non-blocking mode, the fix should go into os.File.Read.

Patch-set: 4
Reviewer: Gerrit User 5056 <5056@62eb7196-b449-3ce5-99f1-c037f21e1705>
Label: Code-Review=0