net/http: Use error vars to avoid string comparison.
Update: use plain errors.New rather than ProtocolError

This is a change to transport.go that removes simply returning errors.New(...) as the error state and instead returns a public instance of ProtocolError such that consumers of this api no longer need to do a string comparison in order to identify the error that was returned.

ProtocolError provides an Error function, so this reamins backwards compatible with consumers who are doing string comparison.

This is only a first change, this should be applied to all of net/http - this addresses #9383

Change-Id: I60de26a12abfd8fe4029a4ff06d2da373f009865