Update patch set 1
Patch Set 1:

The issue relates to the usage of string comparison to determine errors, not the example error that was given. I believe the underlying cause was #3514 as was commented there, but that is unrelated to the issue of fixing up the api so we don't break consumers code if an error message is changed.

Patch-set: 1