database/sql: Reuse existing statements in *Tx.Stmt
Previously *Tx.Stmt always prepared a new statement, even if an existing one
was available on the connection the transaction was on.  Now we first see if
the statement is already available on the connection and only prepare if it
isn't.  Additionally, when we do need to prepare one, we store it in the parent
*Stmt to allow it to be later reused by other calls to *Tx.Stmt on that
statement or just straight up by *Stmt.Exec et al.

To make sure that the statement doesn't disappear unexpectedly, we record a
dependency from the statement returned by *Tx.Stmt to the *Stmt it came from
and set a new field, parentStmt, to point to the originating *Stmt.  When the
transaction's *Stmt is closed, we remove the dependency.  This way the "parent"
*Stmt can be closed by the user without her having to know whether any
transactions are still using it or not.

Change-Id: I9e823b38293a0973d171999f796758e83dabddf0