net/mail: moved some code to internal/mime
The code concerning quoted-printable encoding (RFC 2045) and its
variant for MIME headers (RFC 2047) is currently spread in
mime/multipart and net/mail. It is also not exported.

This commit is the second step to fix that issue. It moves the
code concerning RFC 2047 from net/mail to internal/mime. The
exposed API is unchanged.

Concerns #4943.

Change-Id: I5f58aa58e74bbe4ec91b2e9b8c81921338053b00