Update patch set 3
Patch Set 3:

If the compiler would not call runtime.uint64div and do magic multiplies for 64 bit here that would help the simple loop version with divide by 10 greatly. It would still make some unnecessary multiplies and shifts but it should not be that big of an issue to still require the special casing. (to be benchmarked) 

If the special casing is currently to much i can change the patch to only do the other changes.

Patch-set: 3