Update patch set 1
Patch Set 1:

> If you're trying to match known elliptic.Curve instances, can't you just compare to the result of elliptic.P256() etc?

To get the same semantics, you'd need to perform a bytewise comparison on the CurveParams. (Consider if a library included an alternative implementation of one of these curves. Cf. the comments re point compression.)

(That also necessarily results in initializing the curves in elliptic/curve, which may not be useful if the curve being compared against isn't one of those.)

> I would expect a Name member to contain a human-readable name which generally wouldn't be suitable for string matching since it might change.

Well, they are "named curves." Curve.CurveName would be stuttering; so Curve.Name seems best.

> (Also, FIPS names rather than SEC names are used everywhere else.)

Sure; that's easy enough.

Patch-set: 1
Label: Code-Review=0