Update patch set 4
Patch Set 4:

(12 comments)

Impressive speedups. Some comments below.

Most notably, the spec doesn't guarantee that a uint is 32bit on a 32bit platform - it's likely, but it could be 64bits.

A more reliable type to use is uintptr which is guaranteed to hold a pointer of the target platform; on a modern CPU that size of a uintptr is likely the size of a general-purpose register.

Patch-set: 4
Label: Code-Review=0