Update patch set 4
Patch Set 4:

Thanks for the very good and helpful comments! I will gladly incorporate them and test the performance.
I choose uint because according to spec its either 32bit or 64bit and in both cases the code is correct. Currently its 32bit on the 32bit platforms that i can test and therefore yields a speedup. If its 64 bit on 32bit it might just be slower. While uintptr seems currently at least 32bit it does not have a guaranteed minimal size of 32bit on all platforms in the spec. So i assumed (even it might be unlikely such architectures are or will be supported) it can be e.g. 16bit or even lower since the spec does not seem to forbid that. If uintptr is guaranteed to always be 32bit or higher of course that point is mood. Maybe the spec already demands this elsewhere or it could be clarified?

Patch-set: 4
Label: Code-Review=0