path/filepath: prevent Join from inadvertently create a UNC path on Windows
Join inadvertently creates a Universal Naming Convention (UNC)[0] path
on Windows when told to join at least three non-empty path elements,
where the first one is `\` or `/`.

Prevent this from happening by creating a `\\\` prefix, which is not
deemed to be a UNC path. The multiple `\` elements in the prefix will
be replaced with a single one when the path is Cleaned.

Updates golang/go#9167.

While here, clean up test by removing unnecessary helper function.

[0]: http://msdn.microsoft.com/en-us/library/gg465305.aspx

Change-Id: Ib6eda597106cb025137673b33c4828df1367f75b