Update patch set 3
Patch Set 3:

> What is the problem with itoa(0)? It yields "0" for me here. It'd
 > be good to add a test for it, whatever it is.

sorry that was indeed an error in my testing. I can replicate the base 10 test cases from strconv.Itoa if needed into net, but thought that might be to much. Instead then it might be an option to consider to use strconv in the few places its needed. But that seems not to be desired in the net package judging by the top comments in parse.go.

Patch-set: 3
Label: Code-Review=0