Update patch set 1
Patch Set 1:

I wrote a program that creates the number of threads specified on command line,
and the memory usage difference for 600 vs. 1200 threads is about 6MB (about
2 pages per thread on 32-bit windows xp). And note the value is definitely smaller
than the specified commit size of 65535 bytes.

I specified stack reservation size to 1MB in the test program, so the increasing the
stack reservation size won't increase the memory usage much if the thread doesn't
actually use that much stack space (which is usually the case).

Patch-set: 1
Label: Code-Review=0