Update patch set 1
Patch Set 1:

Alternatively the handling for negative numbers can be removed and a comment added that this itoa function does not handle negative numbers contrary to any other itoa implementation in the golang code base. Or it can be renamed to uitoa wich takes uint and all calls to it in log need to be wrapped in uint casts.

Patch-set: 1
Label: Code-Review=0