cmd/ld, runtime: increase OS thread stack reservation on Windows
Some DLL functions use large stacks. But Go sets maximum thread size
to 64K. Increase that to 1M on 386 and 2M on amd64 to make it similar
to what most Windows programs do.

Because the reservation size is not the size of stack memory initially
committed, increasing the stack reservation limit won't increase the
(stack) memory usage by itself. And my tests showed that the maximum
number of OS thread that could be created within one process is not
affected by this change.

Fixes #9457.

Change-Id: I3a82f7d9717d3d55211b6eb1c34b00b0eaad83ed