Update patch set 2
Patch Set 2:

I will leave the renaming and tests out of the revised version.

16 seems better due to less work for insertion Sort. Its better if the array is partitionable into equally sized blocks. Now one can argue if thats more likely for 16 or for 20 in general. BlockSort implementations i have read seem to suggest 16 to 31. I would actually test and implement dynamic scaling in a later patch. Do you know where the 20 came from in the first place? 16 yields better results for me. 

Line 375 is not only linear search. Its a one pass linear insertion to put data[a] into the right place into data[a:b] when data[a+1:b] is already sorted. This will need same amount of swaps regardless if we search for the position linear and swap on the way or search binary and swap the range later. However we can test if the less comparisons when search binary and swapRange later is faster. Also putting it past the binary search phase for the new partitioning might be a good idea i have to test.

Patch-set: 2