Update patch set 3
Patch Set 3:

Ok there are some performance gains to be had. Sometimes its minor changes that are hard to benchmark. Changing to blocksize 16 is a major gain due to the sizes of the arrays being powers of 2. I changes the insertion to binary insertion sort. Will benchmark now and change the position of some checks before it can be merged.

Patch-set: 3
Label: Code-Review=0