Update patch set 4
Patch Set 4:

Good catch on the botched benchmark. It dawned on me something was not right yesterday but it was to late and i was already afk.
Will be corrected with the nice version that sorts back and forth.

The checks for reverse and sorted blocks seem to only really matter when the blocks are small or the data is in the respective order. Otherwise they just add small overhead.

Yes the swap for loop can be a swapRange. Sadly all the swapRange calls are not inlined :( Would be great if we could get them to inline or manually do so if its worth the code smell.

Patch-set: 4
Label: Code-Review=0