Update patch set 13
Patch Set 13:

Thanks for the hints and comments. I understand the simplicity argument and this patch might already add to much code to warrant the speedup. 

Btw. In light of simplicity whats your take on factoring out the code into two functions: binarySearchFirst and binarySearchLast to make the symmerge function cleaner? Will pay a small function call overhead. Therefore we should further special case on b-a == 2 at the beginning to not make calls to these functions when only searching for one element.

The data.Less guarantees are still met. Since its still only O(log n) comparisons but we save O(log n) calls on symmerge. This was also proposed in the paper that introduced symmerge. The difference is just that sometimes if you compare e.g. one element to two others you might get lucky to only need one comparison instead of two. When i debugged it i see the new code being faster sometimes and the other being faster on others. On average it seems the old code has on average better hit rate than the new. 

To be on the save side we can patch (if accepted) this in a version that has exactly the same number of data.Less calls.

Do you think anything speaks against patching the blocksize to 16 first with arguments and benchmarks that in generally this seems to work better and on the tests produces also less data.Less calls?

I have resorted to testing on a bare minimum FreeBSD box with a G3220 dual core, no turbo boost, all c states disabled that i can, minimal environment running. I still get +-2% :(

Thanks for hinting at the best option. I will use that in the future.

Patch-set: 13