Update patch set 4
Patch Set 4:

> Old algorithm copies all alive objects. It cause 2400byte~ memcpy with bound check.

I don't see why it would have to copy, nor do I see how the new algorithm is supposed to avoid that.

To further prove my point, the new algorithm has a bug in it: the last line should slice only up to j, not i.

But in the end, I don't see the need to try and squeeze every cycle out of this particular loop.  We only execute it if there have been more than N connection deaths, and N doesn't need to be 1.  We could even reuse dead slots at the end of connStmt() to make it even less important to optimize that loop.

Patch-set: 4
Label: Code-Review=0