Update patch set 13
Patch Set 13:

I will do some benchmarks with the -best option.

I hope i understood your question right:
The binary searches are not symmetric because they need to put the element left or right of a range of equal values. e.g. if we have  [2] merge [1 2 2 2 3] and we want to find the position for 2 we need to put it here [1 _ 2 2 2 3] to be stable. if we have [1 2 2 2 3] [2] we need to put it here [1 2 2 2 _ 3] to keep stability. Other implementations of similar sorts name it similar to binarySearchLeft and binarySearchRight. Does that answer the question? 

We can also make it exactly behave like the original version. The first search would then be:

minpos := m
maxpos := b-1
for minpos < maxpos {
  c := minpos + (maxpos-minpos)/2
  if data.Less(c, a) {
    minpos = c + 1
  } else {
    maxpos = c
  }
}
if minpos == b-1 {
  if data.Less(b-1,a){
    minpos = b
  }
}

Patch-set: 13