runtime: Get rid of all the trailing empty arrays
The ones at the end of M and G are unused.
The one at the end of itab is variable-sized, and at least one.
The ones at the end of interfacetype and uncommontype are not
needed, as the preceding slice references them (the slice was
originally added for use by reflect?).
The one at the end of stackmap is already accessed correctly,
and the runtime never allocates one.

Change-Id: Ia75e3aaee38425f038c506868a17105bd64c712f