crypto/ecdsa: make Sign safe with broken entropy sources
This is a first stab. It is, I think, the minimal (in terms
of patch size) solution:

   - csprng_key = ChopMD-256(SHA2-512(priv.D||entropy||hash))
   - reader = AES-256-CTR(k=csprng_key)

This, however, provides at most 128-bit collision-resistance,
so that Adv will have a term related to the number of messages
signed that is significantly worse than plain ECDSA. Though this
does not seem to be of any practical importance.

(If you like the small size of the patch enough, I can likely get
one of our crypto folks to derive the concrete security result
formally. Unless, of course, one exists that I'm not aware of.)

Perhaps a better alternative is to use an HMAC-DRBG construction,
but this will be a significantly larger patch.

GitHub issue: https://github.com/golang/go/issues/9452

Tests:

  TestNonceSafety: Check that signatures are safe even with a
    broken entropy source.

  TestINDCCA: Check that signatures remain non-deterministic
    with a functional entropy source.

Change-Id: Ie7e04057a3a26e6becb80e845ecb5004bb482745