Update patch set 1
Patch Set 1:

(7 comments)

The use of SHA-512 over SHA-256 and the use of AES seem a little non-parsimonious. However, this is ECDSA and, as such, a pretty high-level primitive already. It's reasonable to assume that users of ECDSA probably already include AES.

So the outstanding question is why SHA-512-256 rather than just SHA-256?

Otherwise, looks good.

(The magic string in the commit message for referencing an issue is "Fixes #9452". That should automatically close the issue when submitted too.)

Patch-set: 1
Label: Code-Review=0