Update patch set 1
Patch Set 1:

I really don't like this kind of solution.

we're bloating the binary for performance gains,
and the gain is only for integers < 1024, which
happens to be used by our benchmarks.

what about small-magnitude negative integers?

Teaching the compiler to allocate the objects
on stack when the interfaces that contains them
don't escape seems like a better generic solution.

Patch-set: 1