Update patch set 1
Patch Set 1:

> > Teaching the compiler to allocate the objects
 > > on stack when the interfaces that contains them
 > > don't escape seems like a better generic solution.
 > 
 > I agree we should allocate on the stack when we can (#8618).  But
 > that's not the only situation where these things can happen. 
 > Returning error codes is one example, the interface does escape (to
 > the parent) but we would nevertheless like to avoid allocation.

But normally errors are not simply integer either.

I feel the change is optimizing for our benchmark, rather than
for the underlying problem. (Like iant said, the major speed-
ups are from the fmt benchmarks, but they happen to use
integer covered in this CL.)

Suppose we change the fmt benchmark to use 1025, although
the benchmark is still correct, it won't benefit from this change
anymore.

It's still early in the cycle, I guess we can postpone optimization
like this and see if we can do better in the compiler.

Patch-set: 1
Reviewer: Gerrit User 5055 <5055@62eb7196-b449-3ce5-99f1-c037f21e1705>
Label: Code-Review=0