runtime: cache stack roots if goroutine is blocked
On one of our production systems, we have thousands of goroutines
being blocked on I/O for long periods of time. Since scanning the
stack is one of the more expensive parts of GC, this amounted to a
large chunk of CPU time.

Since the stack cannot change if a G is blocked, we can cache the
results of the stack scan if the G has been idle since the last
GC. By doing this, we get around having to run through gentraceback,
which is expensive.

This change cuts a significant time off GC. This simple program

	package main

	import (
		"runtime"
	)

	func main() {
		runtime.GCstarttimes(2)
		v := make(chan bool)
		for i := 0; i < 15000; i++ {
			go func() {
				<-v
			}()
		}
		 for {
			v := make([]byte, 2<<20)
			_ = v
		}
	}

goes from scan phase time of 29.403ms to 4.382ms.

Change-Id: I8bdb89bda223afa4a5d1a92f066fa8952a2b1a4c