Update patch set 4
Patch Set 4:

This is here for symmetry, not for performance. I am going to remove the assembly implementation. I do not want to give people any justification for using this function other than that it is the clearest way to write the code.

Patch-set: 4
Reviewer: Gerrit User 5056 <5056@62eb7196-b449-3ce5-99f1-c037f21e1705>
Label: Code-Review=0