runtime: (SUPER HACKY) refactor runtime using adonovan's sockdrawer tool
DO NOT SUBMIT/DO NOT REVIEW

This is an attempted refactoring of the runtime package using
adonovan's sockdrawer tool.  I picked the runtime.clusters file in the
cl produced the runtime package and made changes to try to get the
refactored runtime package to compile.  For simplicity the refactored
runtime package only contains code for darwin/amd64. This should be
easy to fix the refactoring is automated.

The code in this CL is not very pretty.  I did the bare minimum to
satisfy compiler errors.

Current status (update as of 22 Jan 2015) go_bootstrap compiles and works!
But there are a number of functions that still have not been renamed
properly.

The next steps are to first get all.bash to complete successfully, and
then to automate the production of this refactored and hacky
darwin/amd64 runtime.  Most of the changes in this CL were made
manually to scope out the work required by the refactoring.

See https://codereview.appspot.com/186270043/ for the sockdrawer tool.

Change-Id: Icdbe3fa54ce16c78acb6d84490463cc203a76b34