runtime: (SUPER HACKY) refactor runtime with adonovan's sockdrawer
DO NOT SUBMIT/DO NOT REVIEW

This is an attempted refactoring of the runtime package using
adonovan's sockdrawer tool.  I picked the runtime.clusters file in the
cl produced the runtime package and made changes to try to get the
refactored runtime package to compile.  For simplicity the refactored
runtime package only contains code for darwin/amd64. This should be
easy to fix the refactoring is automated.

The code in this CL is not very pretty.  I did the bare minimum to
satisfy compiler errors.

Current status (update as of 23 Jan 2015) cmd/go compiles and runs
but some tests fail. Rewriting of asm_amd64.s has been automated.

The next step is to automate the rest of the production of this
refactored and hacky darwin/amd64 runtime.  Most of the changes in
this CL were made manually to scope out the work required by the
refactoring.

See golang.org/cl/3269 for the tools used to produce the refactoring
and https://codereview.appspot.com/186270043/ for adonovan's original
sockdrawer CL.

Change-Id: Icdbe3fa54ce16c78acb6d84490463cc203a76b34