Update patch set 1
Patch Set 1: Code-Review-1

This may be worth doing, or it may not. We'll need to measure once the rest of the system is actually running with performance in mind (right now we are still focusing on correctness). It is great to have the code available to try, and thank you very much for that, but we need to wait until we're in a position where we can give it a fair trial.

Patch-set: 1