Update patch set 1
Patch Set 1:

> The other piece of this good idea is to scan goroutine stacks as
 > they are preempted by the scheduler and their status is about to be
 > moved out of _Grunning. If we scan on preemption a concurrent GC
 > cycle can avoid rescanning stacks/cached roots of every G unless it
 > has been _Grunning during the cycle.

Rick, can you elaborate on this idea? I don't understand how it will work.
First, you can't do transitive marking before GC has not started. Because you need the writebarrier installed to not get black to white pointers.
Second, assuming that we don't do transitive closure but just collect gray objects pointed by the stack; if a goroutine stack is scanned several times, we need to somehow discard results of all but the last scan. Otherwise we will retain objects that are garbage for a long time; but were marked by one of the first scans of the stack.

Patch-set: 1