net/http: allow to use IPv6 zone identifiers in URIs
This change is a follow-up to change 2431 and requires URL parsers in
change 2431.

Using IPv6 link-local addresses to make connections between on-link
nodes is useful for small distributed applications but it requires zone
identifiers to distinguish a correct IP link. It's the same for
transports using URI for destination discovery such as HTTP, WebSocket.

This change allows HTTP clients, proxies or other intermediaries to use
a literal IPv6 address followed by a zone identifier within a URI as
described in RFC 6874.

Fixes #9544.

Change-Id: I7406bd0aa961d260d96f1f887c2e45854e921452