Update patch set 1
Patch Set 1:

Rick Hudson wrote:  (copying this back to Gerrit from email)

"""
The steps relevant to this discussion.

1. Enter new concurrent GC cycle
2. Scan all G stacks greying direct referents
3. Turn on write barrier
4. Do concurrent transitive mark phase exhausting all wbufs
---- At this point the only source of pointers to white objects are on G stacks
5. Stop the world for mark termination
6. Scan all G stacks transitively walking any white object discovered.
7. Start the world ending this GC cycle

Step 6 is what scanning on preemption hopes to optimize. If a G in 6 has not run since step 2 then we know that it will not hold any references to a white object since step 2 greyed all white objects reachable from its stack. CL 3030 (this CL) is mostly about optimizing scans between separate GC cycles, not about avoiding the rescan of stacks that have not changed between step 2 and step 6. CL 3030 will help here of course but not as much as doing nothing. Looking a bit further down the road to where we make step 6 concurrent this becomes even more important since we will need to scan G's repeatedly as part of the distributed termination algorithm.

During steps 2,3,4,5 the optimization will scan as we move any G from _Grunning. These stacks will no longer refer to white objects and step 4 will exhaust all wbufs generated by such stack scans. This means at step 6 the only G stacks that need to be scanned are the few in _Grunning actively associated with a P. 

I suspect, confirmation TBD, this is why our "garbage" latency numbers are not as good as json and splay. Garbage is the only one that has lots of goroutines that are rescanned at 6. I suspect, confirmation TBD, the motivating example for CL 3030 will also be helped by this optimization.
"""

Patch-set: 1